backend/pyproject.toml
[tool.poetry]
name = "food-recipe-backend"
version = "0.1.0"
description = "FastAPI backend for Food Recipe Generator with semantic search"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
sqlalchemy = "^2.0.23"
alembic = "^1.12.1"
psycopg2-binary = "^2.9.9"
pydantic = "^2.5.0"
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"
# ML dependencies (install separately if needed)
# sentence-transformers = "^2.2.2"
# numpy = "^1.26.0"
# scikit-learn = "^1.5.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
httpx = "^0.25.2"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "app.main:app"
dev = "uvicorn app.main:app --reload --host 0.0.0.0 --port 8000"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-ignore = ['E203', 'W503']

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "sentence_transformers.*",
    "sklearn.*",
    "databases.*",
    "asyncpg.*"
]
ignore_missing_imports = true